+++ Research/joseim2/stop_id_retriever.py
@@ -1,4 +1,5 @@
 import requests
+import key
 
 #  key:'headsign', 0:'route_id', 1:'origin', 2:'destination', 3:'scheduled', 4:'expected', 5:'expected_mins', dictionary with info about current buses with same stop_id
