--- Research/jt37/key.py
@@ -1,4 +1,0 @@
-key = "cde8e82f02cb488daa290433126c745f"
-
-def get_key():
-    return key
